% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalDb.R
\name{addModelToLocalDatabase}
\alias{addModelToLocalDatabase}
\title{\code{addModelToLocalDatabase}}
\usage{
addModelToLocalDatabase(model, taxonomy_id, model_source, version,
  scientific_name = "NULL", common_english_name = "NULL",
  description = "NULL")
}
\arguments{
\item{model}{gmt model.}

\item{taxonomy_id}{tax id of species.}

\item{model_source}{where did you found ypour model.}

\item{version}{represents version of the model.}

\item{scientific_name}{name of the species in literature.}

\item{common_english_name}{name of the species in english literature.}

\item{description}{your description about this model. Can contain metadata about models.}
}
\value{
Return DB queries and status of model adding operation.
}
\description{
\code{addModelToLocalDatabase} adds model to database under presented public key (taxonomi_id, model_source, version).
}
