% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{filter_ontology}
\alias{filter_ontology}
\title{Filter Ontology}
\usage{
filter_ontology(gmt, min_nr_of_elements = NULL, max_nr_of_elements = NULL)
}
\arguments{
\item{gmt}{A data frame which contains the data, imported from a GMT file.}

\item{min_nr_of_elements}{minimum size of term. Default 20 percent from
quantile on term size distribution.}

\item{max_nr_of_elements}{maximum size of term. Default 80 percent from
quantile on term size distribution.}
}
\value{
Return data frame with model from specific location.
}
\description{
Filter ontology to only contain terms between given min. and max. sizes.
}
