% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{geneSet}
\alias{geneSet}
\title{Gene Set Example}
\format{
A data frame with 7 rows and 3 variables:
\describe{
  \item{ontologyId}{Ontology ID of the gene set}
  \item{ontologyName}{Description of the gene set}
  \item{listOfValues}{Genes in the gene set}
}
}
\source{
\url{http://geneontology.org/}
}
\usage{
geneSet
}
\description{
Data frame parsed from a Gene Matrix Transposed (GMT) file containing 7 gene
sets. Each row represents a gene set.
}
\keyword{datasets}
