% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalDb.R
\name{getModelFromLocalDatabaseAsDf}
\alias{getModelFromLocalDatabaseAsDf}
\title{\code{getModelFromLocalDatabaseAsDf}}
\usage{
getModelFromLocalDatabaseAsDf(taxonomy_id, model_source, version)
}
\arguments{
\item{taxonomy_id}{tax id of species.}

\item{model_source}{model source, which you provide adding model to DB.}

\item{version}{represents version of the model.}
}
\value{
Return data frame which include model.
}
\description{
\code{getModelFromLocalDatabaseAsDf} takes copy of the model in data frame form.
}
