% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalDb.R
\name{getModelFromLocalDatabaseAsList}
\alias{getModelFromLocalDatabaseAsList}
\title{\code{getModelFromLocalDatabaseAsList}}
\usage{
getModelFromLocalDatabaseAsList(taxonomy_id, model_source, version)
}
\arguments{
\item{taxonomy_id}{tax id of species.}

\item{model_source}{model source, which you provide adding model to DB.}

\item{version}{represents version of the model.}
}
\value{
Return list which include model.
}
\description{
\code{getModelFromLocalDatabaseAsList} takes copy of the model in list form.
}
