% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{InputOutputFunctions}
\alias{InputOutputFunctions}
\alias{readGmtFileAsDataFrame}
\alias{saveDataFrameAsGmtFile}
\title{Input/Output Functions}
\usage{
readGmtFileAsDataFrame(gmtFilePath)

saveDataFrameAsGmtFile(modelDF, gmtFilePath)
}
\arguments{
\item{gmtFilePath}{path with name of file, where the file is localized. Example: "/R/MulEA/extdata/model.gmt"}

\item{modelDF}{data frame with model.}

\item{gmtFilePath}{path with name of file, where to save model. Example: "R/MulEA/extdata/savedModel.gmt"}
}
\value{
Return data frame with model from specific location.

Return gmt file under specific location which include model in gmt format.
}
\description{
\code{readGmtFileAsDataFrame}

\code{readGmtFileAsDataFrame} read model in data frame form from gmt file.

\code{saveDataFrameAsGmtFile}

\code{saveDataFrameAsGmtFile} saves copy of the model from data frame in gmt file.
}
\examples{
modelDfFromFile <- MulEA::readGmtFileAsDataFrame(gmtFilePath = system.file(package="MulEA", "extdata", "model.gmt"))
modelDfFromFile <- MulEA::readGmtFileAsDataFrame(gmtFilePath = system.file(package="MulEA", "extdata", "model.gmt"))
MulEA::saveDataFrameAsGmtFile(modelDF = modelDfFromFile, gmtFilePath = "~/fromDb.gmt")
}
