% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypergeometricTest.R
\docType{class}
\name{MuleaHypergeometricTest-class}
\alias{MuleaHypergeometricTest-class}
\alias{MuleaHypergeometricTest}
\alias{runTest,MuleaHypergeometricTest-method}
\title{PRIVATE class : An S4 class to represent a Hypergeometric tests in Mulea.}
\usage{
\S4method{runTest}{MuleaHypergeometricTest}(testObject)
}
\arguments{
\item{testObject}{Object of s4 class represents Mulea Test.}
}
\description{
PRIVATE class : An S4 class to represent a Hypergeometric tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{runTest}: runs test calculations.
}}

\section{Slots}{

\describe{
\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{testData}}{A data from expeciment to analize accross model.}

\item{\code{pool}}{A background data to count test.}
}}

