% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Plot Barplot}
\usage{
plot_barplot(
  reshaped_results,
  selected_rows_to_plot = NULL,
  ontology_id_colname = "ontologyId",
  p_value_type_colname = "adjustedPValue",
  p_value_max_threshold = 0.05
)
}
\arguments{
\item{reshaped_results}{data.table in relaxed form.}

\item{selected_rows_to_plot}{numeric; which rows of the reshaped results data
frame should be included in the plot?}

\item{ontology_id_colname}{character}

\item{p_value_type_colname}{character}

\item{p_caue_max_threshold}{numeric}
}
\value{
Return a barplot.
}
\description{
Plots barplot of p-values.
}
\seealso{
\code{\link{reshape_results}}
}
