% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot Graph (Network)}
\usage{
plot_graph(
  reshaped_results,
  shared_elements_min_threshold = 0,
  p_value_type_colname = "adjustedPValue",
  ontology_id_colname = "ontologyId",
  ontology_element_colname = "genIdInOntology",
  p_value_max_threshold = 0.05
)
}
\arguments{
\item{reshaped_results}{data.table in relaxed form.}

\item{shared_elements_min_threshold}{numeric}

\item{p_value_type_colname}{character}

\item{ontology_id_colname}{character}

\item{ontology_element_colname}{numeric}

\item{p_value_max_threshold}{numeric}
}
\value{
Return a graph.
}
\description{
Plots graph representation of enrichment results.
}
\seealso{
\code{\link{reshape_results}}
}
