% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Plot Lollipop}
\usage{
plot_lollipop(
  reshaped_results,
  ontology_id_colname = "ontology_id",
  selected_rows_to_plot = NULL,
  p_value_type_colname = "eFDR",
  p_value_max_threshold = 0.05
)
}
\arguments{
\item{reshaped_results}{data.table in relaxed form, obtained as the output of the `reshape_results` function. The data source for generating the barplot.}

\item{ontology_id_colname}{Character, specifies the column name that contains ontology IDs in the input data.}

\item{selected_rows_to_plot}{A numeric vector specifying which rows of the reshaped results data frame should be included in the plot. Default is NULL.
frame should be included in the plot?}

\item{p_value_type_colname}{Character, specifies the column name for p-values in the input data. Default is 'eFDR'.}

\item{p_value_max_threshold}{Numeric, representing the maximum p-value threshold for filtering data. Default is 0.05.}
}
\value{
Returns a lollipop plot
}
\description{
Plots lollipop plot of p-values.
}
\details{
Create a customized  lollipop plot of p-values, facilitating visual exploration and analysis of statistical significance within ontology categories.
}
\examples{
# import example gene set
# import other gene sets from a GMT file using read_gmt()
data(geneSet) 
Run model on geneset
ora_model <- ora(
 gmt = geneSet,
 element_names = selectDf$select, 
 background_element_names = poolDf$background_element_names,
 p_value_adjustment_method = "eFDR",
 number_of_permutations = 1000
)
ora_results <- run_test(ora_model)
Reshape results
ora_reshaped_results <- reshape_results(
 model = ora_model, 
 model_results = ora_results, 
 p_value_type_colname='eFDR'
)
plot_lollipop(
reshaped_results = ora_reshaped_results,
p_value_max_threshold=0.05,
p_value_type_colname = "eFDR"
)

}
\seealso{
\code{\link{reshape_results}}
}
