% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{poolDf}
\alias{poolDf}
\title{TITLE}
\format{
A data frame with 33 rows and 2 variables:
\describe{
  \item{X}{Row names}
  \item{background_element_names}{DEFINITION}
}
}
\source{
SOURCE
}
\usage{
poolDf
}
\description{
DESCRIPTION
}
\keyword{datasets}
