% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankedBasedTest.R
\docType{class}
\name{RankedBasedTest-class}
\alias{RankedBasedTest-class}
\alias{RankedBasedTest}
\alias{runTest,RankedBasedTest-method}
\title{An S4 class to represent a ranked based tests in Mulea.}
\usage{
\S4method{runTest}{RankedBasedTest}(testObject)
}
\arguments{
\item{testObject}{Object of s4 class represents Mulea Test.}
}
\description{
An S4 class to represent a ranked based tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{runTest}: Run test calculations.
}}

\section{Slots}{

\describe{
\item{\code{method}}{A method from ranked methods to count results.}

\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{testData}}{A data from expeciment to analize accross model.}

\item{\code{scores}}{A vectore of scores per testData.}

\item{\code{p}}{A power of weight.}

\item{\code{numberOfPermutations}}{A number of permutations used in renked test.}
}}

