% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{read_gmt}
\alias{read_gmt}
\title{Read GMT File}
\usage{
read_gmt(file)
}
\arguments{
\item{file}{character; a path which points to a GMT file.}
}
\value{
Returns data frame with the model from a specific location.
}
\description{
Read model data from a Gene Matrix Transposed (GMT) file and parse into a
data frame.
}
