% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{readGmtFileAsDF}
\alias{readGmtFileAsDF}
\title{\code{readGmtFileAsDF}}
\usage{
readGmtFileAsDF(gmtFilePath)
}
\arguments{
\item{gmtFilePath}{path with name of file, where the file is localized. Example: "/home/mulea/files/lastModel.gmt"}
}
\value{
Return data frame with model from specific location.
}
\description{
\code{readGmtFileAsDF} read model in data frame form from gmt file.
}
\examples{
muleaPkgDir <- find.package("MulEA")
modelDfFromFile <- MulEA::readGmtFileAsDF(gmtFilePath = paste(muleaPkgDir,"/example/model.gmt", sep = ""))
}
