% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{saveModelFromDataFrameToGmtFile}
\alias{saveModelFromDataFrameToGmtFile}
\title{\code{saveModelFromDataFrameToGmtFile}}
\usage{
saveModelFromDataFrameToGmtFile(modelDF, gmtFilePath)
}
\arguments{
\item{modelDF}{data frame with model.}

\item{gmtFilePath}{path with name of file, where to save model. Example: "/hmoe/mulea/files/lastModel.gmt"}
}
\value{
Return gmt file under specific location which include model in gmt format.
}
\description{
\code{saveModelFromDataFrameToGmtFile} saves copy of the model from data frame in gmt file.
}
