% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalDb.R
\name{saveModelFromLocalDatabaseToFile}
\alias{saveModelFromLocalDatabaseToFile}
\title{\code{saveModelFromLocalDatabaseToFile}}
\usage{
saveModelFromLocalDatabaseToFile(taxonomy_id, model_source, version,
  gmtFilePath)
}
\arguments{
\item{taxonomy_id}{tax id of species.}

\item{model_source}{model source, which you provide adding model to DB.}

\item{version}{represents version of the model.}

\item{gmtFilePath}{path with name of file, where to save model. Example: "/hmoe/mulea/files/lastModel.gmt"}
}
\value{
Return gmt file under specific location which include model in gmt format.
}
\description{
\code{saveModelFromLocalDatabaseToFile} saves copy of the model in file.
}
