% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetBasedEnrichmentTest.R
\docType{class}
\name{SetBasedEnrichmentTest-class}
\alias{SetBasedEnrichmentTest-class}
\alias{SetBasedEnrichmentTest}
\alias{run_test,SetBasedEnrichmentTest-method}
\title{PRIVATE class : An S4 class to represent a Hypergeometric tests in Mulea.}
\usage{
\S4method{run_test}{SetBasedEnrichmentTest}(model)
}
\arguments{
\item{model}{Object of s4 class represents Mulea Test.}
}
\value{
SetBasedEnrichmentTest object. Used as private function.

run_test method for SetBasedEnrichmentTest object. Used as private function.
}
\description{
PRIVATE class : An S4 class to represent a Hypergeometric tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run_test(SetBasedEnrichmentTest)}: runs test calculations.

}}
\section{Slots}{

\describe{
\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{element_names}}{A data from expeciment to analize accross model.}

\item{\code{pool}}{A background data to count test.}

\item{\code{nthreads}}{Number of processor's threads used in calculations.}
}}

\examples{
\dontrun{
#It is a private s4 object. Look at ora's examples.
}
\dontrun{
#It is a private method. Look at run_test of ora's examples.
}
}
