% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetBasedTest.R
\docType{class}
\name{SetBasedTest-class}
\alias{SetBasedTest-class}
\alias{SetBasedTest}
\title{An S4 class to represent a set based tests in Mulea.}
\description{
An S4 class to represent a set based tests in Mulea.
}
\section{Slots}{

\describe{
\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{testData}}{A data from expeciment to analize accross model.}

\item{\code{pool}}{A background data co count test.}

\item{\code{adjustMethod}}{A type of algorithm used to adjust values.}
}}

