% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalDb.R
\name{startLocalDatabase}
\alias{startLocalDatabase}
\title{\code{startLocalDatabase}}
\usage{
startLocalDatabase(muleaDBLocalization = ":memory:")
}
\arguments{
\item{muleaDBLocalization}{Character which represents path to directory where should be created MulEA.sqlite file.
If not provided DB will be created in memory. Example of path: "/home/mulea/countingdb/"}
}
\value{
If all inputs are integer and logical, then the output will be an DB creation status.
}
\description{
\code{startLocalDatabase} creates database, which can helps in working with big amount of models.
DB has to be created to make queries on it.
}
\examples{
creationOfLocalDB <- MulEA::startLocalDatabase(":memory:")
}
