% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalDb.R
\name{stopLocalDatabase}
\alias{stopLocalDatabase}
\title{\code{stopLocalDatabase}}
\usage{
stopLocalDatabase()
}
\value{
Result is the status of process of DB closing.
}
\description{
\code{stopLocalDatabase} stops Mulea DB. It is important whean you have in plans to work for a long time with DB.
It helps with data consystency.
}
\examples{
creationOfLocalDB <- MulEA::startLocalDatabase(":memory:")
stopDbResults <- MulEA::stopLocalDatabase()
}
