% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubramanianTest.R
\docType{class}
\name{SubramanianTest-class}
\alias{SubramanianTest-class}
\alias{SubramanianTest}
\alias{runTest,SubramanianTest-method}
\title{PRIVATE class : An S4 class to represent a ranked based tests in Mulea.}
\usage{
\S4method{runTest}{SubramanianTest}(testObject)
}
\arguments{
\item{testObject}{Object of s4 class represents Mulea Test.}
}
\value{
KolmogorovSmirnovTest object. Used as private object.

runTest method for SubramanianTest object. Used as private function.
}
\description{
PRIVATE class : An S4 class to represent a ranked based tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{runTest}: runs test calculations.
}}

\section{Slots}{

\describe{
\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{testData}}{A data from expeciment to analize accross model.}

\item{\code{scores}}{A vectore of scores per testData.}

\item{\code{p}}{A power of weight.}

\item{\code{numberOfPermutations}}{A number of permutations used in renked test.}
}}

\examples{
\dontrun{
#It is a private s4 object. Look at RankedBasedTest's examples.
}
\dontrun{
#It is a private method. Look at runTest of RankedBasedTest's examples.
}
}
