% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankedBasedTest.R
\docType{class}
\name{GSEA-class}
\alias{GSEA-class}
\alias{GSEA}
\alias{run_test,GSEA-method}
\title{An S4 class to represent a ranked based tests in Mulea.}
\usage{
\S4method{run_test}{GSEA}(model)
}
\arguments{
\item{model}{Object of s4 class represents Mulea Test.}
}
\value{
GSEA object. This object represents ranked based tests in
Mulea.

run_test method for GSEA object. Returns results of
counting using methods from ranking based area.
}
\description{
An S4 class to represent a ranked based tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run_test}: runs test calculations.
}}

\section{Slots}{

\describe{
\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{testData}}{A data from expeciment to analize accross model.}

\item{\code{element_scores}}{A vectore of element_scores per testData.}

\item{\code{p}}{A power of weight. Default value is 1.}

\item{\code{element_score_type}}{Defines the GSEA score type. Only positive element_scores - "pos",
only negative element_scores - "neg" and mixed (standard) - "std".}

\item{\code{number_of_permutations}}{A number of permutations used in KS test. Default
vlue is 1000.}
}}

\examples{
modelDfFromFile <- MulEA::readGmtFileAsDataFrame(
  gmtFilePath = system.file(package="MulEA", "extdata", "model.gmt"))
dataFromExperiment <- c("FBgn0004407", "FBgn0010438", "FBgn0003742",
                        "FBgn0029709", "FBgn0030341", "FBgn0037044",
                        "FBgn0002887", "FBgn0028434", "FBgn0030170",
                        "FBgn0263831")
dataFromExperimentScores <- c(0.09, 0.11, 0.15, 0.20, 0.21, 0.24, 0.28, 0.30,
                              0.45, 0.50)
GSEASubramanian <- GSEA(gmt = modelDfFromFile,
                                              testData = dataFromExperiment,
                                              element_scores = dataFromExperimentScores)
GSEASubramanianRes <- MulEA::run_test(GSEASubramanian)
}
