% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankedBasedTest.R
\docType{class}
\name{gsea-class}
\alias{gsea-class}
\alias{gsea}
\alias{run_test,gsea-method}
\title{Gene Set Enrichment Analysis (GSEA)}
\usage{
\S4method{run_test}{gsea}(model)
}
\arguments{
\item{model}{Object of s4 class represents Mulea Test.}
}
\value{
GSEA object. This object represents ranked based tests.

run_test method for GSEA object. Returns results of
counting using methods from ranking based area.
}
\description{
An S4 class to represent a ranked based tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run_test(gsea)}: runs test calculations.

}}
\section{Slots}{

\describe{
\item{\code{gmt}}{A data frame which contains the data, imported from a GMT file.}

\item{\code{element_names}}{A vector of elements names to include in the analysis,
ordered by their scores.}

\item{\code{element_scores}}{A vector of element_scores per element_names.}

\item{\code{gsea_power}}{A power of weight. Default value is 1.}

\item{\code{element_score_type}}{Defines the GSEA score type. Only positive
element_scores - "pos", only negative element_scores - "neg" and mixed
(standard) - "std".}

\item{\code{number_of_permutations}}{The number of permutations used in KS test.
Default value is 1000.}

\item{\code{test}}{character}
}}

\examples{
modelDfFromFile <- MulEA::read_gmt(
  file = system.file(package="MulEA", "extdata", "model.gmt"))
dataFromExperiment <- c("FBgn0004407", "FBgn0010438", "FBgn0003742",
                        "FBgn0029709", "FBgn0030341", "FBgn0037044",
                        "FBgn0002887", "FBgn0028434", "FBgn0030170",
                        "FBgn0263831")
dataFromExperimentScores <- c(0.09, 0.11, 0.15, 0.20, 0.21, 0.24, 0.28, 0.30,
                              0.45, 0.50)
GSEASubramanian <- gsea(gmt = modelDfFromFile,
                        element_names = dataFromExperiment,
                        element_scores = dataFromExperimentScores)
GSEASubramanianRes <- MulEA::run_test(GSEASubramanian)
}
