% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R, R/Utils.R
\name{InputOutputFunctions}
\alias{InputOutputFunctions}
\alias{relaxModelAndResults}
\alias{plotGraph}
\alias{plotBarplot}
\alias{readGmtFileAsDataFrame}
\alias{saveDataFrameAsGmtFile}
\title{Input/Output Functions}
\usage{
relaxModelAndResults(mulea_model = NULL, mulea_model_resuts = NULL)

plotGraph(mulea_relaxed_resuts, edge_weight_cutoff = 0)

plotBarplot(
  result_data_frame,
  selection_vector = 1:10,
  categories_names = "DB_names",
  probabilities_values = "FDR"
)

readGmtFileAsDataFrame(gmtFilePath)

saveDataFrameAsGmtFile(modelDF, gmtFilePath)
}
\arguments{
\item{mulea_model}{MulEA object represents model. For example created by MulEA::ORA.}

\item{mulea_model_resuts}{Results from model, in most cases it is returned by MulEA::runTest generic method.}

\item{mulea_relaxed_resuts}{data.table in relaxed form.}

\item{result_data_frame}{data.frame returned from test.}

\item{selection_vector}{}

\item{gmtFilePath}{path with name of file, where the file is localized or where to save model. Example: "/R/MulEA/extdata/model.gmt"}

\item{modelDF}{data frame with model.}
}
\value{
Return relaxed datatable where model and results are merged for plotting purposes.

Return plot.

Return plot.

Return data frame with model from specific location.

Return gmt file under specific location which include model in gmt format.
}
\description{
\code{relaxModelAndResults}

\code{relaxModelAndResults} merge model and model relsuts into 
one relaxed datatable for easy resutls graphical interpretation.

\code{plotGraph}

\code{plotGraph} merge model and model relsuts into 
one relaxed datatable for easy resutls graphical interpretation.

\code{plotBarplot}

\code{plotBarplot} barplot of p-values.

\code{readGmtFileAsDataFrame}

\code{readGmtFileAsDataFrame} read model in data frame form from gmt file.

\code{saveDataFrameAsGmtFile}

\code{saveDataFrameAsGmtFile} saves copy of the model from data frame in gmt file.
}
\examples{
modelDfFromFile <- MulEA::readGmtFileAsDataFrame(gmtFilePath = system.file(package="MulEA", "extdata", "model.gmt"))
MulEA::saveDataFrameAsGmtFile(modelDF = modelDfFromFile, gmtFilePath = paste(system.file(package="MulEA", "extdata"), "fromDb.gmt", sep = "/"))
}
