% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R, R/Utils.R
\name{InputOutputFunctions}
\alias{InputOutputFunctions}
\alias{reshape_results}
\alias{plot_graph}
\alias{plot_barplot}
\alias{plot_heatmap}
\alias{read_gmt}
\alias{write_gmt}
\alias{filter_ontology}
\alias{list_to_gmt}
\title{Input/Output Functions}
\usage{
reshape_results(
  model = NULL,
  model_results = NULL,
  model_ontology_col_name = "ontologyId",
  ontology_id_colname = "ontology_id",
  p_value_type_colname = "eFDR",
  p_value_max_threshold = TRUE
)

plot_graph(
  reshaped_results,
  ontology_id_colname = "ontology_id",
  ontology_element_colname = "element_id_in_ontology",
  shared_elements_min_threshold = 0,
  p_value_type_colname = "eFDR",
  p_value_max_threshold = 0.05
)

plot_barplot(
  reshaped_results,
  ontology_id_colname = "ontology_id",
  selected_rows_to_plot = NULL,
  p_value_type_colname = "eFDR",
  p_value_max_threshold = 0.05
)

plot_heatmap(
  reshaped_results,
  ontology_id_colname = "ontology_id",
  ontology_element_colname = "element_id_in_ontology",
  p_value_type_colname = "eFDR",
  p_value_max_threshold = 0.05
)

read_gmt(file)

write_gmt(gmt, file)

filter_ontology(gmt, min_nr_of_elements = NULL, max_nr_of_elements = NULL)

list_to_gmt(gmt_list)
}
\arguments{
\item{model}{a MulEA model, created e.g. by ora().}

\item{model_results}{Results from model, returned by run_test().}

\item{model_ontology_col_name}{character}

\item{ontology_id_colname}{character}

\item{p_value_type_colname}{character}

\item{p_value_max_threshold}{logical}

\item{reshaped_results}{data.table in relaxed form.}

\item{ontology_element_colname}{numeric}

\item{shared_elements_min_threshold}{numeric}

\item{selected_rows_to_plot}{vector for selecting variables to plot.}

\item{file}{path with name of file, where the file is localized or where to save model. Example: "/R/MulEA/extdata/model.gmt"}

\item{gmt}{input dataframe, read from gmt file.}

\item{min_nr_of_elements}{minimum size of term. Default 20 percent from quantile on term size distribution.}

\item{max_nr_of_elements}{maximum size of term. Default 80 percent from quantile on term size distribution.}

\item{gmt_list}{input list with elements names as ontologyId and genes in each element.}
}
\value{
Return detailed and relaxed datatable where model and results are
merged for plotting purposes.

Return plot.

Return plot.

Return plot.

Returns data frame with the model from a specific location.

Returns the model as a .gmt file at a specific location.

Return data frame with model from specific location.

Return data frame with model.
}
\description{
\code{reshape_results} merge model and model results into
one relaxed datatable.

\code{plot_graph} Plots graph representation of enrichment results.

\code{plot_barplot} Plots barplot of p-values.

\code{plot_heatmap} Plots heatmap of enriched terms and obtained p-values.

\code{read_gmt} read model in data frame form from gmt file.

\code{write_gmt} saves copy of the model from dataframe as a .gmt file.

\code{filter_ontology} Filters ontology to only contain terms between given min. and max. sizes.

\code{list_to_gmt}

\code{list_to_gmt} conver ontology representation from list to gmt dataframe.
}
\examples{
modelDfFromFile <- read_gmt(
  file = system.file(package="MulEA", "extdata", "model.gmt"))
write_gmt(gmt = modelDfFromFile,
          file = paste(system.file(package="MulEA", "extdata"),
          "fromDb.gmt", sep = "/"))
}
