% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R, R/Utils.R
\name{InputOutputFunctions}
\alias{InputOutputFunctions}
\alias{reshapeResults}
\alias{plotGraph}
\alias{plotBarplot}
\alias{plotHeatmap}
\alias{readGmtFileAsDataFrame}
\alias{saveDataFrameAsGmtFile}
\alias{filterOntology}
\alias{convertListToGmtDataFrame}
\title{Input/Output Functions}
\usage{
reshapeResults(
  mulea_model = NULL,
  mulea_model_resuts = NULL,
  mulea_model_ontology_col_name = "ontologyId",
  mulea_model_resuts_ontology_col_name = "ontologyId",
  category_stat_column_name = "adjustedPValue",
  cut_off_to_test_data = TRUE
)

plotGraph(
  mulea_relaxed_resuts,
  edge_weight_cutoff = 0,
  statistics_value_colname = "adjustedPValue",
  ontology_id_column_name = "ontologyId",
  gen_id_in_ontology_column_name = "genIdInOntology",
  statistics_value_cutoff = 0.05
)

plotBarplot(
  mulea_relaxed_resuts,
  selection_vector = NULL,
  categories_names = "ontologyId",
  statistics_value_colname = "adjustedPValue",
  statistics_value_cutoff = 0.05
)

plotHeatmap(
  mulea_relaxed_resuts,
  statistics_value_colname = "adjustedPValue",
  gen_id_in_ontology_column_name = "genIdInOntology",
  statistics_value_cutoff = 0.05
)

readGmtFileAsDataFrame(gmtFilePath)

saveDataFrameAsGmtFile(modelDF, gmtFilePath)

filterOntology(input_gmt, min = NULL, max = NULL)

convertListToGmtDataFrame(ontologyReprAsList)
}
\arguments{
\item{mulea_model}{the MulEA object represents model. For example created by
MulEA::ORA.}

\item{mulea_model_resuts}{Results from model, in most cases returned by
MulEA::run_test generic method.}

\item{mulea_relaxed_resuts}{data.table in relaxed form.}

\item{selection_vector}{vector for selecting variables to plot.}

\item{gmtFilePath}{path with name of file, where the file is localized or where to save model. Example: "/R/MulEA/extdata/model.gmt"}

\item{modelDF}{data frame with model.}

\item{input_gmt}{input dataframe, read from gmt file.}

\item{min}{minimum size of term. Default 20 percent from quantile on term size distribution.}

\item{max}{maximum size of term. Default 80 percent from quantile on term size distribution.}

\item{ontologyReprAsList}{input list with elements names as ontologyId and genes in each element.}
}
\value{
Return detailed and relaxed datatable where model and results are
merged for plotting purposes.

Return plot.

Return plot.

Return plot.

Returns data frame with the model from a specific location.

Returns the model as a .gmt file at a specific location.

Return data frame with model from specific location.

Return data frame with model.
}
\description{
\code{reshapeResults} merge model and model results into
one relaxed datatable.

\code{plotGraph} Plots graph representation of enrichment results.

\code{plotBarplot} Plots barplot of p-values.

\code{plotHeatmap} Plots heatmap of enriched terms and obtained p-values.

\code{readGmtFileAsDataFrame} read model in data frame form from gmt file.

\code{saveDataFrameAsGmtFile} saves copy of the model from dataframe as a .gmt file.

\code{filterOntology} Filters ontology to only contain terms between given min. and max. sizes.

\code{convertListToGmtDataFrame}

\code{convertListToGmtDataFrame} conver ontology representation from list to gmt dataframe.
}
\examples{
modelDfFromFile <- MulEA::readGmtFileAsDataFrame(gmtFilePath = system.file(package="MulEA", "extdata", "model.gmt"))
MulEA::saveDataFrameAsGmtFile(modelDF = modelDfFromFile, gmtFilePath = paste(system.file(package="MulEA", "extdata"), "fromDb.gmt", sep = "/"))
}
