% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R, R/Utils.R
\name{InputOutputFunctions}
\alias{InputOutputFunctions}
\alias{reshape_results}
\alias{plot_graph}
\alias{plot_barplot}
\alias{plot_heatmap}
\alias{readGmtFileAsDataFrame}
\alias{saveDataFrameAsGmtFile}
\alias{filterOntology}
\alias{convertListToGmtDataFrame}
\title{Input/Output Functions}
\usage{
reshape_results(
  model = NULL,
  model_results = NULL,
  model_ontology_col_name = "ontologyId",
  ontology_id_colname = "ontologyId",
  p_value_type_colname = "adjustedPValue",
  p_value_max_threshold = TRUE
)

plot_graph(
  reshaped_results,
  shared_elements_min_threshold = 0,
  p_value_type_colname = "adjustedPValue",
  ontology_id_colname = "ontologyId",
  ontology_element_colname = "genIdInOntology",
  p_value_max_threshold = 0.05
)

plot_barplot(
  reshaped_results,
  selected_rows_to_plot = NULL,
  ontology_id_colname = "ontologyId",
  p_value_type_colname = "adjustedPValue",
  p_value_max_threshold = 0.05
)

plot_heatmap(
  reshaped_results,
  p_value_type_colname = "adjustedPValue",
  ontology_element_colname = "genIdInOntology",
  p_value_max_threshold = 0.05
)

readGmtFileAsDataFrame(gmtFilePath)

saveDataFrameAsGmtFile(modelDF, gmtFilePath)

filterOntology(input_gmt, min = NULL, max = NULL)

convertListToGmtDataFrame(ontologyReprAsList)
}
\arguments{
\item{model}{the MulEA object represents model. For example created by
MulEA::ora.}

\item{model_results}{Results from model, in most cases returned by
MulEA::run_test generic method.}

\item{reshaped_results}{data.table in relaxed form.}

\item{selected_rows_to_plot}{vector for selecting variables to plot.}

\item{gmtFilePath}{path with name of file, where the file is localized or where to save model. Example: "/R/MulEA/extdata/model.gmt"}

\item{modelDF}{data frame with model.}

\item{input_gmt}{input dataframe, read from gmt file.}

\item{min}{minimum size of term. Default 20 percent from quantile on term size distribution.}

\item{max}{maximum size of term. Default 80 percent from quantile on term size distribution.}

\item{ontologyReprAsList}{input list with elements names as ontologyId and genes in each element.}
}
\value{
Return detailed and relaxed datatable where model and results are
merged for plotting purposes.

Return plot.

Return plot.

Return plot.

Returns data frame with the model from a specific location.

Returns the model as a .gmt file at a specific location.

Return data frame with model from specific location.

Return data frame with model.
}
\description{
\code{reshape_results} merge model and model results into
one relaxed datatable.

\code{plot_graph} Plots graph representation of enrichment results.

\code{plot_barplot} Plots barplot of p-values.

\code{plot_heatmap} Plots heatmap of enriched terms and obtained p-values.

\code{readGmtFileAsDataFrame} read model in data frame form from gmt file.

\code{saveDataFrameAsGmtFile} saves copy of the model from dataframe as a .gmt file.

\code{filterOntology} Filters ontology to only contain terms between given min. and max. sizes.

\code{convertListToGmtDataFrame}

\code{convertListToGmtDataFrame} conver ontology representation from list to gmt dataframe.
}
\examples{
modelDfFromFile <- MulEA::readGmtFileAsDataFrame(gmtFilePath = system.file(package="MulEA", "extdata", "model.gmt"))
MulEA::saveDataFrameAsGmtFile(modelDF = modelDfFromFile, gmtFilePath = paste(system.file(package="MulEA", "extdata"), "fromDb.gmt", sep = "/"))
}
