% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{list_to_gmt}
\alias{list_to_gmt}
\title{Convert list to GMT data frame}
\usage{
list_to_gmt(gmt_list)
}
\arguments{
\item{gmt_list}{List with element names as ontology_id and genes in each
element.}
}
\value{
Return data frame with model.
}
\description{
Convert ontology representation from list to gmt data frame.
}
