% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypergeometricTest.R
\docType{class}
\name{MuleaHypergeometricTest-class}
\alias{MuleaHypergeometricTest-class}
\alias{MuleaHypergeometricTest}
\alias{run_test,MuleaHypergeometricTest-method}
\title{PRIVATE class : An S4 class to represent a Hypergeometric tests in Mulea.}
\usage{
\S4method{run_test}{MuleaHypergeometricTest}(model)
}
\arguments{
\item{model}{Object of s4 class represents Mulea Test.}
}
\value{
MuleaHypergeometricTest object. Used as private function.

run_test method for MuleaHypergeometricTest object. Used as private
function.
}
\description{
PRIVATE class : An S4 class to represent a Hypergeometric tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run_test(MuleaHypergeometricTest)}: runs test calculations.

}}
\section{Slots}{

\describe{
\item{\code{gmt}}{A data.frame representing the GMT model.}

\item{\code{element_names}}{Data to be analysed across the model.}

\item{\code{background_element_names}}{Background data used for the test.}
}}

\examples{
\dontrun{
#It is a private s4 object. Look at ora's examples.
}
\dontrun{
#It is a private method. Look at run_test of ora's examples.
}
}
