% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA.R
\docType{class}
\name{ora-class}
\alias{ora-class}
\alias{ora}
\alias{run_test,ora-method}
\title{An S4 class to represent a set based tests in Mulea.}
\usage{
\S4method{run_test}{ora}(model)
}
\arguments{
\item{model}{Object of s4 class represents Mulea Test.}
}
\value{
ora object. This object represents set based tests.

run_test method for ora object. Returns results of counting using
methods from set based area.
}
\description{
An S4 class to represent a set based tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run_test(ora)}: runs test calculations.

}}
\section{Slots}{

\describe{
\item{\code{method}}{A method from set based methods to count results. Possible
values: "Hypergeometric", "SetBasedEnrichment".}

\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{element_names}}{A data from experiment to analize accross model.}

\item{\code{background_element_names}}{A background data to count test.}

\item{\code{p_value_adjustment_method}}{A type of algorithm used to adjust values.
Possible values: "eFDR" and all from p.adjust {stats} documentation.}

\item{\code{number_of_permutations}}{A number of permutations used in set based
enrichment test. Default value is 10000.}

\item{\code{number_of_cpu_threads}}{Number of processor threads used in calculations.}
}}

\examples{
modelDfFromFile <- read_gmt(
  file = system.file(package="MulEA", "extdata", "model.gmt"))
dataFromExperiment <- c(
  "FBgn0004407", "FBgn0010438", "FBgn0003742", "FBgn0029709", "FBgn0030341",
  "FBgn0037044", "FBgn0002887", "FBgn0028434", "FBgn0030170", "FBgn0263831")
dataFromExperimentPool <- unique(c(
  c("FBgn0033690", "FBgn0261618", "FBgn0004407", "FBgn0010438", "FBgn0032154",
    "FBgn0039930", "FBgn0040268", "FBgn0013674", "FBgn0037008", "FBgn0003116",
    "FBgn0037743", "FBgn0035401", "FBgn0037044", "FBgn0051005", "FBgn0026737",
    "FBgn0026751", "FBgn0038704", "FBgn0002887", "FBgn0028434", "FBgn0030170",
    "FBgn0263831", "FBgn0000579"),
  c("FBgn0066666", "FBgn0000000", "FBgn0099999", "FBgn0011111", "FBgn0022222",
    "FBgn0777777", "FBgn0333333", "FBgn0003742", "FBgn0029709",
    "FBgn0030341")))
setBasedTest <- ora(gmt = modelDfFromFile,
                    element_names = dataFromExperiment, 
                    number_of_cpu_threads = 2)
setBasedTestWithPool <- ora(gmt = modelDfFromFile,
                            element_names = dataFromExperiment,
                           background_element_names = dataFromExperimentPool,
                           number_of_cpu_threads = 2)
setBasedTestWithPoolAndAdjust <- ora(
  gmt = modelDfFromFile,
  element_names = dataFromExperiment,
  background_element_names = dataFromExperimentPool,
  p_value_adjustment_method = "BH",
  number_of_cpu_threads = 2
 )
setBaseTestWithPermutationTestAdjustment <- ora(
  gmt = modelDfFromFile,
  element_names = dataFromExperiment,
  p_value_adjustment_method = "eFDR",
  number_of_cpu_threads = 2
 )
modelDfFromFile <- read_gmt(
  file = system.file(package="MulEA", "extdata", "model.gmt"))
dataFromExperiment <- c(
  "FBgn0004407", "FBgn0010438", "FBgn0003742", "FBgn0029709", "FBgn0030341",
  "FBgn0037044", "FBgn0002887", "FBgn0028434", "FBgn0030170", "FBgn0263831")
dataFromExperimentPool <- unique(c(
  c("FBgn0033690", "FBgn0261618", "FBgn0004407", "FBgn0010438", "FBgn0032154",
    "FBgn0039930", "FBgn0040268", "FBgn0013674", "FBgn0037008", "FBgn0003116",
    "FBgn0037743", "FBgn0035401", "FBgn0037044", "FBgn0051005", "FBgn0026737",
    "FBgn0026751", "FBgn0038704", "FBgn0002887", "FBgn0028434", "FBgn0030170",
    "FBgn0263831", "FBgn0000579"),
  c("FBgn0066666", "FBgn0000000", "FBgn0099999", "FBgn0011111", "FBgn0022222",
    "FBgn0777777", "FBgn0333333", "FBgn0003742", "FBgn0029709",
    "FBgn0030341")))
setBasedTest <- ora(
  gmt = modelDfFromFile,
  element_names = dataFromExperiment,
  number_of_cpu_threads = 2
 )
setBasedTestWithPool <- ora(
  gmt = modelDfFromFile,
  element_names = dataFromExperiment, 
  background_element_names = dataFromExperimentPool,
  number_of_cpu_threads = 2
)
setBasedTestWithPoolAndAdjust <- ora(
  gmt = modelDfFromFile,
  element_names = dataFromExperiment,
  background_element_names = dataFromExperimentPool,
  p_value_adjustment_method = "BH", number_of_cpu_threads = 2
)
setBaseTestWithPermutationTestAdjustment <- ora(
  gmt = modelDfFromFile,
  element_names = dataFromExperiment,
  p_value_adjustment_method = "eFDR",
  number_of_cpu_threads = 2
)
setBasedTestRes <- run_test(setBasedTest)
setBasedTestWithPoolRes <- run_test(setBasedTestWithPool)
setBasedTestWithPoolAndAdjustRes <- run_test(setBasedTestWithPoolAndAdjust)
setBaseTestWithPermutationTestAdjustmentRes <- run_test(setBaseTestWithPermutationTestAdjustment)
}
