% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(
  reshaped_results,
  p_value_type_colname = "adjustedPValue",
  ontology_element_colname = "genIdInOntology",
  p_value_max_threshold = 0.05
)
}
\arguments{
\item{reshaped_results}{data.table in relaxed form.}

\item{p_value_type_colname}{character}

\item{ontology_element_colname}{character}

\item{p_value_max_threshold}{numeric}
}
\value{
Return a heatmap.
}
\description{
Plots heatmap of enriched terms and obtained p-values.
}
\seealso{
\code{\link{reshape_results}}
}
