% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankedBasedTest.R
\docType{class}
\name{RankedBasedTest-class}
\alias{RankedBasedTest-class}
\alias{RankedBasedTest}
\alias{runTest,RankedBasedTest-method}
\title{An S4 class to represent a ranked based tests in Mulea.}
\usage{
\S4method{runTest}{RankedBasedTest}(testObject)
}
\arguments{
\item{testObject}{Object of s4 class represents Mulea Test.}
}
\value{
RankedBasedTest object. This object represents ranked based tests in Mulea.

runTest method for RankedBasedTest object. Returns results of counting using methods from ranking based area.
}
\description{
An S4 class to represent a ranked based tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{runTest}: runs test calculations.
}}

\section{Slots}{

\describe{
\item{\code{method}}{A method from ranked methods to count results.}

\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{testData}}{A data from expeciment to analize accross model.}

\item{\code{scores}}{A vectore of scores per testData.}

\item{\code{p}}{A power of weight. Default vlue is 1.}

\item{\code{numberOfPermutations}}{A number of permutations used in renked test. Default vlue is 1000.}
}}

\examples{
muleaPkgDir <- find.package("MulEA")
modelDfFromFile <- MulEA::readGmtFileAsDF(gmtFilePath = paste(muleaPkgDir,"/example/model.gmt", sep = ""))
dataFromExperiment <- c("FBgn0004407", "FBgn0010438", "FBgn0003742", "FBgn0029709", "FBgn0030341", "FBgn0037044", "FBgn0002887", "FBgn0028434", "FBgn0030170", "FBgn0263831")
dataFromExperimentScores <- c(0.09, 0.11, 0.15, 0.20, 0.21, 0.24, 0.28, 0.30, 0.45, 0.50)
rankedBasedTestKs <- RankedBasedTest(method = "KS", gmt = modelDfFromFile, testData = dataFromExperiment)
rankedBasedTestSubramanian <- RankedBasedTest(method = "Subramanian", gmt = modelDfFromFile, testData = dataFromExperiment, scores = dataFromExperimentScores)
rankedBasedTestKsRes <- MulEA::runTest(rankedBasedTestKs)
rankedBasedTestSubramanianRes <- MulEA::runTest(rankedBasedTestSubramanian)
}
