% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{read_gmt}
\alias{read_gmt}
\title{Read GMT File}
\usage{
read_gmt(file)
}
\arguments{
\item{file}{character; a path which points to a GMT file.}
}
\value{
Returns a data frame with three columns: 
ontology_id: The identifier for each entry.
ontology_name: The name or description corresponding to the ontology_id.
list_of_values: A list of values associated with each ontology_id.
}
\description{
Read model data from a Gene Matrix Transposed (GMT) file and parse into a
data frame.
}
\examples{
# import example gene set
library(mulea)
tf_gmt <- read_gmt(file = system.file(package="mulea", "extdata", "Transcription_factor_RegulonDB_Escherichia_coli_GeneSymbol.gmt"))
}
