% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalDb.R
\name{removeModelFromLocalDatabase}
\alias{removeModelFromLocalDatabase}
\title{\code{removeModelFromLocalDatabase}}
\usage{
removeModelFromLocalDatabase(taxonomy_id, model_source, version)
}
\arguments{
\item{taxonomy_id}{tax id of species.}

\item{model_source}{model source, which you provide adding model to DB.}

\item{version}{represents version of the model.}
}
\value{
Return status of delete operation.
}
\description{
\code{removeModelFromLocalDatabase} removes model from DB.
}
\examples{
muleaPkgDir <- find.package("MulEA")
modelDfFromFile <- MulEA::readGmtFileAsDF(gmtFilePath = paste(muleaPkgDir,"/example/model.gmt", sep = ""))
creationOfLocalDB <- MulEA::startLocalDatabase(":memory:")
MulEA::addModelToLocalDatabase(model = modelDfFromFile, taxonomy_id = 9001, model_source = "GO", version = 0)
MulEA::removeModelFromLocalDatabase(taxonomy_id = 9001, model_source = "GO", version = 0)
stopDbResults <- MulEA::stopLocalDatabase()
}
