% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{reshape_results}
\alias{reshape_results}
\title{Reshape Results}
\usage{
reshape_results(
  model = NULL,
  model_results = NULL,
  model_ontology_col_name = "ontologyId",
  ontology_id_colname = "ontologyId",
  p_value_type_colname = "adjustedPValue",
  p_value_max_threshold = TRUE
)
}
\arguments{
\item{model}{a MulEA model, created e.g. by ora().}

\item{model_results}{Results from model, returned by run_test().}

\item{model_ontology_col_name}{character}

\item{ontology_id_colname}{character}

\item{p_value_type_colname}{character}

\item{p_value_max_threshold}{logical}
}
\value{
Return detailed and relaxed datatable where model and results are
merged for plotting purposes.
}
\description{
This function merges model and model results into a single data frame.
}
\seealso{
\code{\link{plot_graph}}, \code{\link{plot_barplot}},
\code{\link{plot_heatmap}}
}
