% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{saveModelFromDataFrameToGmtFile}
\alias{saveModelFromDataFrameToGmtFile}
\title{\code{saveModelFromDataFrameToGmtFile}}
\usage{
saveModelFromDataFrameToGmtFile(modelDF, gmtFilePath)
}
\arguments{
\item{modelDF}{data frame with model.}

\item{gmtFilePath}{path with name of file, where to save model. Example: "/hmoe/mulea/files/lastModel.gmt"}
}
\value{
Return gmt file under specific location which include model in gmt format.
}
\description{
\code{saveModelFromDataFrameToGmtFile} saves copy of the model from data frame in gmt file.
}
\examples{
muleaPkgDir <- find.package("MulEA")
modelDfFromFile <- MulEA::readGmtFileAsDF(gmtFilePath = paste(muleaPkgDir,"/example/model.gmt", sep = ""))
creationOfLocalDB <- MulEA::startLocalDatabase(":memory:")
MulEA::addModelToLocalDatabase(model = modelDfFromFile, taxonomy_id = 9001, model_source = "GO", version = 0)
modelDfFromLocalDB <- MulEA::getModelFromLocalDatabaseAsDf(taxonomy_id = 9001, model_source = "GO", version = 0)
MulEA::saveModelFromDataFrameToGmtFile(modelDF = modelDfFromLocalDB, gmtFilePath = "~/fromDb.gmt")
stopDbResults <- MulEA::stopLocalDatabase()
}
