% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{selectDf}
\alias{selectDf}
\title{TITLE}
\format{
A data frame with 13 rows and 3 variables:
\describe{
  \item{X}{Row names}
  \item{select}{DEFINITION}
  \item{score}{DEFINITION}
}
}
\source{
SOURCE
}
\usage{
selectDf
}
\description{
DESCRIPTION
}
\keyword{datasets}
