% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetBasedEnrichmentTest.R
\docType{class}
\name{SetBasedEnrichmentTest-class}
\alias{SetBasedEnrichmentTest-class}
\alias{SetBasedEnrichmentTest}
\title{PRIVATE class : An S4 class to represent a Hypergeometric tests in Mulea.}
\value{
SetBasedEnrichmentTest object. Used as private function.
}
\description{
PRIVATE class : An S4 class to represent a Hypergeometric tests in Mulea.
}
\section{Slots}{

\describe{
\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{testData}}{A data from expeciment to analize accross model.}

\item{\code{pool}}{A background data to count test.}
}}

\examples{
\dontrun{
#It is a private s4 object. Look at SetBasedTest's examples.
}
}
