% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubramanianTest.R
\docType{class}
\name{SubramanianTest-class}
\alias{SubramanianTest-class}
\alias{SubramanianTest}
\alias{run_test,SubramanianTest-method}
\title{PRIVATE class : An S4 class to represent a ranked based tests in Mulea.}
\usage{
\S4method{run_test}{SubramanianTest}(model)
}
\arguments{
\item{model}{Object of s4 class represents Mulea Test.}
}
\value{
dataframe with presented columns 'ontology_id', 'ontology_name',
'nrCommonGenesOntologySet', 'nrCommonGenesOntologyBackground',
'pValue', 'adjustedPValue'

run_test method for SubramanianTest object. Used as private function.
}
\description{
PRIVATE class : An S4 class to represent a ranked based tests in Mulea.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run_test(SubramanianTest)}: runs test calculations.

}}
\section{Slots}{

\describe{
\item{\code{gmt}}{A data.frame representing GMT's reprezentation of model.}

\item{\code{element_names}}{A data from expeciment to analize accross model.}

\item{\code{element_scores}}{A vectore of element_scores per element_names.}

\item{\code{p}}{A power of weight.}

\item{\code{element_score_type}}{Defines the GSEA score type. Only positive element_scores - "pos", only negative element_scores - "neg" and mixed (standard) - "std".}
}}

\examples{
\dontrun{
#It is a private s4 object. Look at GSAE's examples.
}
\dontrun{
#It is a private method. Look at run_test of GSEA's examples.
}
}
