% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{write_gmt}
\alias{write_gmt}
\title{Write GMT file}
\usage{
write_gmt(gmt, file)
}
\arguments{
\item{gmt}{A data frame containing the data to be written, imported from a GMT file.}

\item{file}{character; a path naming the output file.}
}
\value{
Returns the model as a .gmt file at a specific location.
}
\description{
Write GMT takes a data frame with specific formatting (columns representing ontology IDs, descriptions, and associated lists of values) 
and writes it to a file in a standardized GMT format.
}
\examples{
#' library(mulea)
tf_gmt <- read_gmt("Transcription_factor_RegulonDB_Escherichia_coli_GeneSymbol.gmt")
tf_gmt_filtered <- filter_ontology(gmt = tf_gmt,
                          min_nr_of_elements = 3,
                          max_nr_of_elements = 400)
write_gmt(gmt = tf_gmt_filtered, file = "Transcription_factor_RegulonDB_Escherichia_coli_GeneSymbol_filtered.gmt")
}
