% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{write_gmt}
\alias{write_gmt}
\title{Write GMT file}
\usage{
write_gmt(gmt, file)
}
\arguments{
\item{gmt}{A data frame which contains the data, imported from a GMT file.}

\item{file}{character; a path naming a file.}
}
\value{
Returns the model as a .gmt file at a specific location.
}
\description{
Write the model data to a Gene Matrix Transposed (GMT) file.
}
\examples{
modelDfFromFile <- read_gmt(
  file = system.file(package="MulEA", "extdata", "model.gmt"))
write_gmt(gmt = modelDfFromFile,
          file = paste(system.file(package="MulEA", "extdata"),
          "fromDb.gmt", sep = "/"))
}
